IMU display now shows orientation and angular velocity.
